builder/requirements.txt
# Runtime dependencies
google-cloud-storage==2.17.0
pyyaml>=6.0.1
colorlog>=6.8.0
psutil>=6.1.0
requests==2.31.0
tqdm==4.66.2
packaging==23.2
setuptools>=70.0.0

# Pin cryptographic dependencies to secure versions
cryptography>=42.0.0
urllib3>=2.0.7